requirements.txt
Django==1.6.1
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-geoposition==0.1.5
django-jquery==1.9.1
django-storages==1.1.8
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
static==1.0.2
wsgiref==0.1.2
